pyproject.toml
[tool.poetry]
name = "techstack-api"
version = "0.1.0"
description = "A Django REST API that serves as the backend for the Techstack blog app."
authors = ["David Semke <david@semke.ca>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
python-decouple = "^3.8"
dj-database-url = "^2.3.0"
rich = "^14.0.0"
requests = "^2.32.3"
psycopg = {extras = ["binary"], version = "^3.2.7"}
django-debug-toolbar = "^5.2.0"
djangorestframework-simplejwt = "^5.5.0"
djoser = "^2.3.1"
django-filter = "^25.1"
django-silk = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
